Renamed classes to match conventions in the Contributions file. Updated cmake files to include new directory structure under algorithms.
