+++ src/utilities/eicrecon/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.16)
 
-print_header(">>>>   E X E C U T A B L E :   algorithms    <<<<")       # Fancy printing
+print_header(">>>>   E X E C U T A B L E :   eicrecon    <<<<")       # Fancy printing
 
 project(eicrecon_project)
